Move inner misc classes up one level
