+++ org.hl7.fhir.convertors/src/main/java/org/hl7/fhir/convertors/misc/argonaut/Context.java
@@ -1,0 +1,19 @@
+package org.hl7.fhir.convertors.misc.argonaut;
+
+import lombok.Data;
+import org.hl7.fhir.dstu3.model.Coding;
+import org.hl7.fhir.dstu3.model.DateTimeType;
+import org.hl7.fhir.dstu3.model.Encounter;
+import org.hl7.fhir.dstu3.model.Reference;
+
+@Data
+public class Context {
+  private String baseId;
+  private Reference authorRef;
+  private Encounter encounter;
+  private Coding encClass;
+  private int obsId;
+  private DateTimeType now = DateTimeType.now();
+  private int orgId;
+  private Reference subjectRef;
+}
