+++ org.hl7.fhir.convertors/src/main/java/org/hl7/fhir/convertors/misc/argonaut/Stats.java
@@ -1,0 +1,10 @@
+package org.hl7.fhir.convertors.misc.argonaut;
+
+import lombok.Data;
+
+@Data
+public class Stats {
+  private int instances;
+  private int errors;
+  private int warnings;
+}
