+++ org.hl7.fhir.convertors/src/main/java/org/hl7/fhir/convertors/misc/searchparam/SearchParameterProcessor.java
@@ -30,7 +30,6 @@ package org.hl7.fhir.convertors.misc.searchparam;
  */
 
 
-import lombok.Data;
 import org.hl7.fhir.exceptions.FHIRException;
 import org.hl7.fhir.exceptions.FHIRFormatError;
 import org.hl7.fhir.r5.formats.IParser.OutputStyle;
@@ -277,27 +276,6 @@ public class SearchParameterProcessor {
       }
     }
     System.out.println("Map loaded - " + list.size() + " entries");
-  }
-
- @Data
-  public class SPRelationship {
-    private final String r4;
-    private final String r3;
-    private final String r2b;
-    private final String r2;
-
-    public String getByCode(String code) {
-      if ("R4".equals(code))
-        return r4;
-      if ("R3".equals(code))
-        return r3;
-      if ("R2b".equals(code))
-        return r2b;
-      if ("R2".equals(code))
-        return getR2();
-      return null;
-    }
-
   }
